some minor refactor to this amazing code written by @LygoBot
