+++ compile.sh
@@ -1,0 +1,1 @@
+gcc main.C -o battleship
