+++ readme.md
@@ -1,0 +1,2 @@
+# Battleship
+Use the bash script `compile.sh` to compile the program
